Added defintions for procfs_dir_necessary and all necessary functions and variables in the heirarchy
Added struct procfs_dir_entry **cur_entry
Added defintion for procfs_dir_refresh to make all dir entries invalid, get the timestamp
and call procfs_fill_root_dir or update_dir_entries depending on whether the directory
to be refreshed is a root node.
Added definition for function make_dir_invalid by setting dir_entry->valid to 0 for all
the entries
Added a part of the defintion for procfs_fill_root_dir for getting all pids, this is not
completed and a part is left as stub to decide on how to proceed.
Added function header update_dir_entries and has been as left as stub
Added defintion for update_entries_list which creates a linked list of dir entries for
the given directory and set those entries to be valid, also puts stat and other information
for those entries including the hash values
Added defintion for lookup_entry to search for a NODE in a given directory by comparing
using its hash values which are computed by calling hash function.
Added definition for hash function, based on ftpfs code
Added definition for rehash function to extend the hashtable length and copy the existing
values to the new extended hash table.
Added definition for insert function which dir_entry to the hash table