Removed double initializations of global variables and pointers and call to bootstrap root node
Due to a small negligency there was redundant code which used to initialize
all the global variables and pointers twice, these were removed from procfs_init
function.
A call to the function procfs_dir_null_lookup which is used to bootstrapping the
root node was added to procfs_create function.