Clean up error checking
There was an additional if (! err) check before calling set_field_value which
meant that if there was an error in setting one of the previous fields, the err
value would still not be 0, so it caused not to enter the block that sets the
values for other fields which resulted in illegitimate values for all the
further fields.

Further an additional else has been added for if (! err) after the call of
set_field_value, so that if setting of the field fails corresponding values
in the /proc/<PID>/stat file to appear 0 instead of some illegitimate values.